Add DockerfileNodemon Use Docker hub nodemon image to deploy Rename ledger-operator to ledger Move js files into src folder
