+++ deploy/DockerfileNodemon
@@ -1,0 +1,2 @@
+FROM node:alpine
+RUN yarn global add nodemon
