--- ledger-operator/.env
@@ -1,1 +1,0 @@
-REF_NODE_URL=http://localhost:3000/track
