+++ ledger/src/miner.js
